finish requestItem backend

+++ business/src/main/java/eg/iti/shareit/model/dto/ActivityDto.java
@@ -7,6 +7,7 @@
 package eg.iti.shareit.model.dto;
 
 import eg.iti.shareit.common.dto.GenericDto;
+import java.math.BigDecimal;
 import java.util.Date;
 
 /**
@@ -15,7 +16,7 @@ import java.util.Date;
  */
 public class ActivityDto implements java.io.Serializable, GenericDto {
 
-    private int id;
+    private BigDecimal id;
     private ItemDto item;
     private UserDto fromUser;
     private UserDto toUser;
@@ -28,11 +29,11 @@ public class ActivityDto implements java.io.Serializable, GenericDto {
     public ActivityDto() {
     }
 
-    public int getId() {
+    public BigDecimal getId() {
         return id;
     }
 
-    public void setId(int id) {
+    public void setId(BigDecimal id) {
         this.id = id;
     }
